requirements.txt
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
blis==0.7.3
boto3==1.16.25
botocore==1.19.25
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
fastai==2.1.7
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
ipykernel==5.3.4
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.4
nbconvert==5.6.1
numpy==1.19.4
packaging==20.7
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
s3transfer==0.3.3
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.94
six==1.15.0
smmap==3.0.4
spacy==2.3.4
srsly==1.0.4
streamlit==0.71.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.3
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torchvision==0.8.1
tornado==6.1
tqdm==4.54.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
validators==0.18.1
wasabi==0.8.0
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1